build: change tsconfig settings to nodenext
This change updates our tsconfig to use settings consistent with our build setup.  If we're not going to bundle our build code, then we shouldn't be using `bundler` `moduleResolution`.  By switching to `nodenext` it enforces the requirement to have extensions as part of our import paths.

Alternatively, we could switch to bundling our builds, which would be my preference, which case, we'd leave it as `esnext` / `bundler`